+++ src/index.tsx
@@ -14,6 +14,7 @@ import {applyPolyfills, defineCustomElements as colorDefineCustomElements, JSX a
 import {defineCustomElements as chartsDefineCustomElements} from '@deckdeckgo/charts/dist/loader';
 
 import {DetailedHTMLProps, HTMLAttributes} from 'react';
+import {setupConfig} from '@ionic/react';
 
 type StencilProps<T> = {
   [P in keyof T]?: Omit<T[P], 'ref'> | HTMLAttributes<T>;
@@ -35,6 +36,10 @@ declare global {
 applyPolyfills().then(async () => {
   await colorDefineCustomElements(window);
   await chartsDefineCustomElements(window);
+});
+
+setupConfig({
+  backButtonText: '',
 });
 
 // Safari 14 workaround 🤮