+++ src/pages/invoices/Invoices.tsx
@@ -32,10 +32,8 @@ import {contrast} from '../../utils/utils.color';
 
 import {Settings} from '../../models/settings';
 
-import Header from '../../components/header/Header';
-
 import InvoiceModal from '../../modals/invoice/InvoiceModal';
-import {BackupInvoices} from '../../components/backup/invoices/BackupInvoices';
+import {BackupInvoices} from '../../components/backup/BackupInvoices';
 
 const Invoices: React.FC = () => {
   const {t} = useTranslation(['invoices', 'common']);
@@ -52,8 +50,6 @@ const Invoices: React.FC = () => {
   return (
     <IonPage>
       <IonContent>
-        <Header></Header>
-
         <main className="ion-padding">
           <IonHeader>
             <IonToolbar className="title">