+++ src/pages/statistics/Statistics.tsx
@@ -1,18 +1,13 @@
 import React, {useState} from 'react';
 
-import {IonContent, IonHeader, IonToolbar, IonPage, useIonViewDidLeave, useIonViewWillEnter} from '@ionic/react';
-
-import {useTranslation} from 'react-i18next';
+import {IonContent, IonHeader, IonToolbar, IonPage, useIonViewDidLeave, useIonViewWillEnter, IonBackButton, IonButtons} from '@ionic/react';
 
 import {rootConnector} from '../../store/thunks/index.thunks';
 
-import Header from '../../components/header/Header';
 import WeekCharts from '../../components/charts/week/WeekCharts';
 import Summary from '../../components/summary/Summary';
 
 const Statistics: React.FC = () => {
-  const {t} = useTranslation('summary');
-
   const [entered, setEntered] = useState<boolean>(false);
 
   useIonViewWillEnter(() => {
@@ -26,12 +21,12 @@ const Statistics: React.FC = () => {
   return (
     <IonPage>
       <IonContent>
-        <Header></Header>
-
         <main className="ion-padding">
           <IonHeader>
             <IonToolbar className="title">
-              <h1>{t('title')}</h1>
+              <IonButtons slot="start">
+                <IonBackButton defaultHref="/more" />
+              </IonButtons>
             </IonToolbar>
           </IonHeader>